Refactor, thread, and update Innovation to support reSIDfp
The commit refactors the Innovation code into a class structure
similar to the FluidSynth, MT-32, and GUS implementations, which
makes it more ammeable to managing it using our thread-and-queue
approach.

Threading the SID emulation allows us always use the highest quality
emulation routines in realtime, even on modest hardware such as an
800 MHz PC or Raspberry Pi.

Threading the playback was an all-or-nothing change, so unfortunately
intermediate commits were not possible.

It also makes the following changes:
 - The code wasn't compiling as DOSBox-X doesn't use the same
   configuration interface as DOSBox; so this was updated.

 - Moves handling of the 'conf' settings (both setting and getting)
   local to the module.

 - Renames the source files, class, and configuration settings
   to use from "innova" to "innovation" because:

   1. There is no historic evidence that people or games called this
      card an "innova". Game installers called it 'Innovation'.

   2. Given we don't call the Sound Blaster "soundbla" or the
      UltraSound "ultrasou", thus we shouldn't call the Innovation
      "innova".