Add pdcurses to the Windows list of vcpkg's in README.md
This commit adds `pdcurses` to vcpkg's list in README.md
given it's a dependency to build DOSBox's optional debugger.

The logic being that our vcpkg install line should provide
all the dependencies needed to build the user-side features.

Given the debugger is no different than other optional
user-side features, we should therefore include it as a
dependency. It's also neither risky nor resource intensive
to do so.

I made the concious decide /not/ to include the various
ncurses-devel libs on the Linux and macOS side because we
consider the debugger feature fully broken and unusable
on those platforms.